Added Phosphor
Phosphor is a Minecraft mod which works to optimize one of game's most inefficient areas-- the lighting engine. It works on both the client and server, and can be installed on servers without requiring clients to also have the mod.

With Phosphor, the amount of time the game takes to generate new chunks can be greatly reduced, and you can expect to see some of the nastier frame stutters caused by loading chunks on the client simply vanish.

It's a no-compromises solution for improving performance either in single-player or large multi-player servers, and changes no features or behaviors of the vanilla game. In fact, our continued insistence towards not changing vanilla behaviours has even led to the mod being officially allowed in Minecraft speedrunning.